Revert "Fix FHMM output error #493 issue due to newer version of pandas. Tested with pandas 0.19.2."
This reverts commit 417ce3ebe8fc03e222c27b797fc2c872aab9edd3.

+++ nilmtk/disaggregate/combinatorial_optimisation.py
@@ -177,7 +177,7 @@ class CombinatorialOptimisation(Disaggregator):
                 output_datastore.append(key, df)
 
             # Copy mains data to disag output
-            mains_df = pd.DataFrame(chunk, columns=cols, dtype='float32')
+            mains_df = pd.DataFrame(chunk, columns=cols)
             output_datastore.append(key=mains_data_location, value=mains_df)
 
         if data_is_available:
