+++ nilmtk/disaggregate/fhmm_exact.py
@@ -407,7 +407,7 @@ class FHMM(Disaggregator):
 
             # Copy mains data to disag output
             output_datastore.append(key=mains_data_location,
-                                    value=pd.DataFrame(chunk, columns=cols, dtype='float32'))
+                                    value=pd.DataFrame(chunk, columns=cols))
 
         if data_is_available:
             self._save_metadata_for_disaggregation(
